+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 - Added error handling for circular references.
 ## Version 1.0.6
 - Further refined and fixed all functions.
+## Version 1.0.7
+- Fully migrated `arrayEquality` to an iterative implementation.
+- Introduced the `EqualityOptions` and the ability to specify an epsilon for floating-point numbers.